changes optimization
