--- lib/utils/const.dart
@@ -1,31 +1,0 @@
-import 'package:flutter/material.dart';
-
-var mainBlue = Colors.blue;
-var mainOrange = Colors.orange;
-var mainGrey = Colors.grey;
-var smallTextColor = Colors.grey[900];
-var smallTextSize = 14.0;
-var smallTextWeight = FontWeight.w400;
-var midiumTextColor = Colors.grey;
-var midiumTextSize = 18.0;
-var midiumTextWeight = FontWeight.w700;
-var hintTextColor = Colors.grey;
-var hintTextSize = 16.0;
-var hintTextWeight = FontWeight.w400;
-
-var appBarTitleColor = Colors.grey[900];
-
-var mainButtonTextSize = 16.0;
-var mainButtonTextColor = Color.fromRGBO(242, 242, 242, 1);
-var mainBottonTextWeight = FontWeight.w700;
-
-var appBarIconColor = Colors.orange;
-
-var defaultPaddingProcent = 0.1;
-
-var introText =
-    "Ornare enim leo enim urna netus quam sit mattis eget quisque eu";
-var passwordRestoreText =
-    "Enter the email address you used during registration and we will send a new password";
-
-var mainBorderColor = Colors.grey;
