+++ lib/widgets/mainButton.dart
@@ -1,3 +1,4 @@
+import 'package:checkincheckout/constants/theme.dart';
 import 'package:flutter/material.dart';
 
 class MainButton extends StatelessWidget {
@@ -13,7 +14,7 @@ class MainButton extends StatelessWidget {
     return FlatButton(
       padding: EdgeInsets.symmetric(
           vertical: MediaQuery.of(context).size.height * 0.02),
-      child: Text(text, style: Theme.of(context).textTheme.button),
+      child: Text(text, style: TextStyle(color: Gray6, fontSize: 18)),
       color: Theme.of(context).buttonColor,
       onPressed: callBack,
       shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
