+++ lib/widgets/payment_plan_item_widget.dart
@@ -1,13 +1,8 @@
+import 'package:checkincheckout/constants/theme.dart';
+import 'package:checkincheckout/pages/signIn_screen.dart';
 import 'package:flutter/material.dart';
 
-import '../../../utils/const.dart';
-
-import '../signIn_screens/signIn_screen.dart';
-
 class PaymentPlanItemWidget extends StatelessWidget {
-  //stay statSess if favorite state will getted from statemanagment
-  //convert to statefull if favorite status will getted from widget,
-
   BuildContext _buildContext;
 
   final String itemTitle;
